+++ src/utils/index.ts
@@ -1,7 +1,7 @@
 import cleanup from "./cleanup"
 import getResolutions from "./getResolutions"
 import measurePDF from "./measurePDF"
-import optimize from "./optimize"
+import optimizePNG from "./optimizePNG"
 import postPingback from "./postPingback"
 import uploadToS3 from "./uploadToS3"
 
@@ -9,7 +9,7 @@ export {
   cleanup,
   getResolutions,
   measurePDF,
-  optimize,
+  optimizePNG,
   postPingback,
   uploadToS3,
 }
